pyproject.toml
[project]
name = "headlines-solution"
version = "0.1.0"
description = "Решение тестового задания."
authors = [
    {name = "Zavkal"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram (>=3.19.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "ruff (>=0.11.2,<0.12.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
